requirements.txt
Django~=4.2.14
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv~=1.0.1
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.0.8
requests~=2.31.0
django-celery-beat==2.7.0




